+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -942,4 +942,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Dawn::Kb::OSVDB_119927
   end
+  it "must have test for OSVDB_119878" do
+    sc = kb.find("OSVDB_119878")
+    sc.should_not   be_nil
+    sc.class.should == Dawn::Kb::OSVDB_119878
+  end
 end